Groups: allow setting owner trust_level at creation.
* When a User creates a Group, they should automatically be a member
* Every Group member needs to have a Trust Level with that Group
* The Trust Level is stored on the Membership model, not the Group

--> Therefore, we need a way to _specify_ a Trust Level at the point
of Group creation, even though the Group model is not expecting a
`trust_level` arg (since it has no trust_level field).

The way we do this is to customize the `create()` method of the
ModelManager for BorrowdGroups, wherein we can do some special handling
of a `trust_level` parameter. That special handling entails popping off
that argument from the kwargs, if present, because the underlying
BorrowdGroup model isn't expecting it, but then make use of Python's
dynamic malleability by simply adding the trust_level back on to
the _instance_ of the Group after its created.

In a later change, this "smuggled" attribute will be used by a signal
handler, which will pass it through to the Membership record created
for the Group owner at the point of Group creation.