Group creators automatically become moderator members.
Add post-save signal handler for Groups which, at the point of creation,
will ensure that the Group's created_by user is made a member with
moderator status.

Note that this entails updating a lot of previous tests, which were
written on the (incorrect) assumption that even Group creators would
need to be manually added to their own Groups - partly because there
wasn't a good way to express Group Trust Level at the point of
creation before.