Add tests for the shortest possible borrowing flow.
Currently only tested (hence possibly only implemented) a "rejected"
flow, i.e. where the item owner turns down the request to borrow
right off the bat.

Again, in deference to the specificities of the workflow execution
use-case, these tests are unusual in that they are explicitly not
subject to Django's usual transaction-based test isolation: the
tests within this class are intended, indeed required, to be run one
after the other, in a specific sequence, broadly reflecting the steps
of the workflow itself. Again, seems appropriate.

In terms of depth or layers of the system under test, these tests
use Django's RequestFactory to test from the View invocation layer;
so, while not as comprehensive as full HTTP client or Selenium-based
tests, they give a good sense of how the system will operate in
practice.

Next step is to wire up the UI with some HTMX or similar.