Group model separation: BorrowdGroup vs auth.models.Group.
This is more groundwork for implementing moderator approval for group
memberships. To do so, this commit separates BorrowdGroup from
auth.models.Group, i.e. separating the concerns of our domain-specific
Group membership on the one hand from access permissions on the other.

BorrowdGroup used to be a direct subclass of auth.models.Group. Now this
is no longer the case. Instead, BorrowdGroup is its own, standalone
model (a regular subclass of django.db.models.Model), and whenever a new
BorrowdGroup is created, a "corresponding" permissions group is created
in parallel. This gives us the ability for a BorrowdUser to be "in" a
BorrowdGroup, from a domain / "user-facing" perspective, *without*
necessarily being part of the corresponding permissions group - which is
exactly the situation we need when a user's group membership is
"pending", or certain other non-"active" statuses.

The tests we had in place for the existing group / item permission
functionality has enabled these changes to be made with a good degree of
confidence - some light manual testing also seemed fine.

Note that, since this a pretty fundamental change in the core auth
models, there is NO MIGRATION PATH to this new state of affairs: you
will have to delete your database and start from scratch. This has also
given us the opportunity to "compact" several migrations for BorrowdUser
and BorrowdGroup, recreating them from scratch at this point.