Integrate `vite` for hot module reloading and packaging.
[`vite`](https://vite.dev) is the current hotness in frontend build
tooling. Hooray for anything written in Rust - efficiency is green :)

This commit _looks_ like it's doing a lot at once, but most of these
changes are just updates to various dependency metadata files, and
bringing in some mostly-default configs for the new tooling.

The biggest change here is that we now require `node` and `npm`, to
avail ourselves of the `vite` goodness.

This commit basically entails:

* Add `django-vite` as a dependency
* Remove `django_browser_reload` as a depenency, as the `vite` dev
  server is now performing this function and more
* Add tailwindcss vite plugin
* Add super simple main.js as entrypoint for vite bundling process
* Add super simple main.css to pull in local tailwind
* Update base templates:
  - No more tailwind from CDNs! Use locally provide from vite
  - Add vite template tags, enabling hot module replacement
* Updates to .gitignore
* Updates to CONTRIBUTING.md, including setup for `nvm` -> `node` ->
  `vite`

Not tackled yet:

* Converting other deps from CDN to locally bundled