Group listing: proper group trust level.
This commit makes it so that one's proper Trust Level is reflected for
each Group on the Group listing page.

Note that, in order to do this, the generic list view and filter used
for that page are now based on the Membership model, _not_ the Group
model itself any longer. This is because the Trust Level is a property
of a User's _relationship_ with the Group, as captured in the Membership
through-model, rather than of the Group itself.

Note also that despite this change, the View class, Filter class and
template naming still refer to "Group" rather than switching everything
to "Membership". This reflects where one would expect to find these
components based on the app UI.